+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Sin publicar
+- Allow optional icon color usage on benefits items
+
 # v1.34.0
 🚀 1.34.0 🚀
 - Add size to MLBusinessMultipleDescriptionView text and images